restructure files for margin program state (#293)
state.rs is getting quite large, so breaking it up
now to reduce future conflicts
